Split out focus state management
