+++ .changeset/metal-ends-lay.md
@@ -1,0 +1,5 @@
+---
+"@macrostrat/mapbox-react": patch
+---
+
+Refactor map easing and focus management components
