+++ packages/mapbox-react/src/index.ts
@@ -5,3 +5,4 @@ export * from "./terrain";
 export * from "./debugger";
 export * from "./focus-state";
 export * from "./layer-managers";
+export * from "./easing";
